добавляет filmController
